+++ src/data/damage-roll/critical-specialization/spear.ts
@@ -41,6 +41,7 @@ export const actions: AssistantAction[] = [
                     rules: [
                         {
                             key: "GrantItem",
+                            // @ts-expect-error
                             onDeleteActions: {
                                 grantee: "restrict",
                             },