+++ src/data/saving-throw/critical-specialization/brawling.ts
@@ -48,6 +48,7 @@ export const actions: AssistantAction[] = [
                     rules: [
                         {
                             key: "GrantItem",
+                            // @ts-expect-error
                             onDeleteActions: {
                                 grantee: "restrict",
                             },