+++ src/data/skill-check/class-feature/swashbuckler/panache.ts
@@ -46,6 +46,7 @@ export const actions: AssistantAction[] = [
                     rules: [
                         {
                             key: "GrantItem",
+                            // @ts-expect-error
                             onDeleteActions: {
                                 grantee: "restrict",
                             },