+++ src/global.d.ts
@@ -1,5 +1,5 @@
 import { GamePF2e, DamageDamageContextFlag } from "foundry-pf2e";
-import { AssistantMessage } from "message.ts";
+import { AssistantSocket } from "socket.ts";
 import { AssistantStorage } from "storage.ts";
 
 declare module "foundry-pf2e" {