+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
 
+### Added
+
+- Grapple (Critical Success, Success, Failure)
+
 ## [6.0.0] - 2025-01-06
 
 ### Added